Remove references to alert_events (#5707)
## Summary
We are not respecting the frequency setting set on alerts. This is due
to a reference to a table that is no longer used, `AlertEvents`. Update
the references to this table to use the respective table based on the
type of event.

## How did you test this change?
Click test:
1) Create a new error alert
  a) set it to send to #chris-errors-test
  b) Set the Alert Frequency to 5 minutes
2) Navigate to the buttons page (`/1/buttons`)
3) Clock on "Throw an Error" button
- [ ] Error is sent to slack channel
4) Click on "Throw an Error" again
- [ ] Error is not sent to slack channel
5) Click on "Throw a randomized Error"
- [ ] Error is sent to slack channel
6) Wait 5 minutes
7) Click on "Throw an Error" again
- [ ] Error is sent to slack channel

<img width="787" alt="Screenshot 2023-06-22 at 1 18 41 PM"
src="https://github.com/highlight/highlight/assets/17744174/33ef49c8-4848-4664-ad7a-fdd1bab47950">

![Screenshot 2023-06-22 at 5 01 12
PM](https://github.com/highlight/highlight/assets/17744174/57a037ce-2cb5-4e47-92fb-733f16ac6ced)

## Are there any deployment considerations?
N/A

---------

Co-authored-by: Spencer Amarantides <spenceramarantides@Spencers-MacBook-Pro.local>