remove useless query loading (#5834)
## Summary

<!--
Ideally, there is an attached GitHub issue that will describe the "why".

If relevant, use this section to call out any additional information
you'd like to _highlight_ to the reviewer.
-->

An optimization was added in #2923. Somewhere along the way, we removed
references to this optimization and we have N number of unnecessary
requests when we first load the error feed

![Screenshot 2023-07-06 at 1 11 40
PM](https://github.com/highlight/highlight/assets/58678/48f3b2e9-3333-4988-97bf-5783dc36f0bc)

Note: the unnecessary requests only happen on initial load. After that,
they are cached in indexdb due to how `preload.ts` works.


## How did you test this change?

<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

* Build passes since there are no application code references to this
query
* I also cleared my indexdb cache locally, confirmed there are no
requests to `GetRecentErrors`, and that the errors pages have no issues.

## Are there any deployment considerations?

<!--
 Backend - Do we need to consider migrations or backfilling data?
-->

N/A