fix OpenSearch indexing when appending both new and old fields in same request (#5823)
## Summary
- a gorm `Create` with a `Returning` clause was not associating the
returned ids with the correct records. e.g. if we attempted to create 10
records but only 5 were inserted, the new ids would be set on the first
5 records regardless of which ones were actually inserted
- instead of relying on the `Returning` clause, get the count of rows
updated and sort the results of the second query by id to get the latest
N rows
<!--
Ideally, there is an attached GitHub issue that will describe the "why".

If relevant, use this section to call out any additional information
you'd like to _highlight_ to the reviewer.
-->

## How did you test this change?
- set a watchpoint and confirmed via debugging
<!--
Frontend - Leave a screencast or a screenshot to visually describe the
changes.
-->

## Are there any deployment considerations?
- will use the `init-opensearch` Make command to reindex the OpenSearch
fields index
<!--
 Backend - Do we need to consider migrations or backfilling data?
-->