Setting up Edge Function test with Cloudflare worker. (#5905)
The Cloudflare SDK does not work in the Next/Vercel Edge runtime. We're
out of luck there, but we can still use the test harness that I built.

The only path forward at this point is to try to reverse-engineer the
[RichiCoder1](https://github.com/RichiCoder1/opentelemetry-sdk-workers)
lib that we're using for Cloudflare and try to strip it down for the
Edge runtime.


![image](https://github.com/highlight/highlight/assets/878947/9e16b2a0-3d3d-4d1c-9480-acb0f5aa0555)

## How to test

Run `yarn e2e:nextjs` to spin up the Next.js testbed at
`http://localhost:3005/`. Click the "Edge Function: *" buttons to create
successful or errorful results.

*******************
@Vadman97 and I figured out that the Node SDK was broken due to some
dependency hoisting limits, so we removed those limits.

I also discovered that we need a `src/instrumentation.ts` file in
addition to `instrumentation.ts`.


![image](https://github.com/highlight/highlight/assets/878947/2afb7b80-b28f-49b7-9f6c-253ea8c32ad9)