Store session insights generated by Harold (#6216)
## Summary

This adds the SessionInsight model, where each row has the generated
insight and the corresponding sessionID.

If a session already has insights generated, the cache is used and no
OpenAI API call is made.

In the future, we might want to allow people to regenerate the session
insights, but this will depend on usage of this feature.

Also includes refining the event filtering. Now we filter out all
Snapshots, Jank, and Performance events when generating insights since
these are more related to metrics compared to actual interactions and
experience of the user.

Also moved the API calls into a resolver function, so the main
SessionInsight function just takes care of whether a session already has
an insight in the db.

## How did you test this change?

Clicked through a few sessions and generated insights, and generating
them again to check that it's using the stored insights.

## Are there any deployment considerations?

This introduces a new table.