Fixes log alert query not saving (#6633)
## Summary

This resolves an issue when defining a log alert query. When a user
inputs a log query the value was being set into a local variable to
update the histogram but was not updating the formStore that was being
used to populate the mutation input variables. This PR should resolve
this issue and allow for query to be sent along to the mutation.

Related Issue: #6630 

## How did you test this change?

Test both locally and in self deployed version

## Are there any deployment considerations?

None, purely frontend changes

## Does this work require review from our design team?

No

---------

Co-authored-by: Anthony Gregis <agregis@dustlabs.com>
Co-authored-by: TheBasedBob <bob@degods.com>