Allow users to pass in sanitize method for XHR/fetch requests (#6892)
## Summary
While Highlight does provide whitelisting to [redact network header and
body](https://www.highlight.io/docs/getting-started/client-sdk/replay-configuration/recording-network-requests-and-responses#redacting-headers-and-bodies),
the current functionality is very limiting. It is a non-trival job to
whitelist all sort of data types and keys for PIIs, tokens, passwords
and so on by hand. It's specially harder when using graphql due to the
single endpoint nature of the protocol.

Instead of having a rigid list of keywords and url rules, I'd like to be
able to modify the entire network data to my liking based on my
application needs and privacy rules. Allow for the client application to
customize the request/response right before sending any data to its
platform.

One can for instance, modify data from request/response headers, body
etc as well as prevent the entire request/response to even be logged.

## How did you test this change?
1) Add in a `requestResponseSanitizer` to the `networkRecording` hash in
the frontend's Highlight options. Example:
```
requestResponseSanitizer: (pair) => {
    if (pair.request.verb === 'POST') {
        console.log('blocking request', pair)
	return null
    } else {
	console.log('allowing request', pair)
	pair.request.headers ||= {}
	pair.request.headers['editTest'] = 'true'
    }

    return pair
}
```
2) Build the SDK and start a Highlight session
3) Click around to get different types of requests
4) Close the session, and let it process
5) View the processed sessions network logs
- [ ] Requests ignored appropriately
- [ ] Requests updated appropriately
- [ ] No known missing requests


https://www.loom.com/share/47ce07c349f14e9aa3fb802b21221167?sid=7153b018-f6ba-46a5-bff8-6c635b269607

## Are there any deployment considerations?
Monitor that Highlight records properly without the
`requestResponseSanitizer` method

## Does this work require review from our design team?
N/A