Node SDK Resource merge (#7070)
It looks like I broke a resource merge a few months back. It didn't
appear to break anything, because we were setting session data on every
span manually.

BUT, with this resource fixed, we can get automatic tracing data
straight from Next.js.

HT to @Vadman97 for finding the bug!